More Interface
Added a Health bar that can fluctuate betwixt 10 and 0, full health and dead respectively. Removed Life Counter from Interface overlay, since we are going with an HP based system. Thus, hitting 0 hit points means death in the level. Added back the movement between levels as it did not save the first time. Not yet implemented, but working on adding an easy system to update player health when dealt damage by an enemy, or grabs a health item that restores 3HP. Working on system to add Nuts and Bolts when encountered in the level.